refactor(App,GameController): GameController 클래스를 사용한 게임 시작기능 추상화
- 기존 gameStart() 메서드의 내부 기능 분리
- 기존 App 클래스의 gameStart() 를 GameController 클래스의 execute로 추상화 수준 변경
- Car 인스턴스에 상태에 대한 serialization 책임을 OutputView ->  GameController 로 변경
- Radom API 관련 함수 분리, utils 폴더로 이동
- Car.js model 폴더로 이동
