+++ src/App.js
@@ -1,4 +1,3 @@
-import { Random } from '@woowacourse/mission-utils';
 import OutputView from './view/OutputView.js';
 import GameController from './GameController.js';
 
@@ -9,36 +8,10 @@ class App {
     this.#controller = new GameController();
 
     await this.#controller.init();
-    const { cars, tryNumber } = this.#controller.getInfo();
+    const { cars } = this.#controller.getInfo();
+    this.#controller.execute();
 
-    this.gameStart(cars, tryNumber);
     this.result(cars);
-  }
-
-  gameStart(cars, tryNumber) {
-    OutputView.racingStartIntro();
-
-    for (let i = 0; i < tryNumber; i += 1) {
-      cars.forEach((car) => {
-        const randomNumber = this.getRandomNumber();
-        this.tryMove(car, randomNumber);
-
-        const { name, distance } = car.getCarInformation();
-        OutputView.printCarState(name, distance);
-      });
-
-      OutputView.printBlankLine();
-    }
-  }
-
-  tryMove(car, randomNumber) {
-    if (randomNumber >= 4) {
-      return car.move();
-    }
-  }
-
-  getRandomNumber() {
-    return Random.pickNumberInRange(0, 9);
   }
 
   result(cars) {
