+++ src/utils/random.js
@@ -1,0 +1,3 @@
+import { Random } from '@woowacourse/mission-utils';
+
+export const getRandomNumberInRange = (start, end) => Random.pickNumberInRange(start, end);
